refactor(agent): introduce TurnEngine; run_tool_call_loop becomes a shim
Phase 2 of the agent-turn unification. Moves the canonical agentic loop body
out of run_tool_call_loop into engine::run_turn_engine, behavior-preserving, and
introduces the first axis seam:

- engine::ToolSource (+ RegistryToolSource) — owns tool advertisement
  (request_specs) and per-call execution (execute_call). The channel/CLI source
  resolves registry+extra under the visibility whitelist and delegates to
  run_one_tool, exactly as before.
- engine::run_turn_engine — the single loop (turn/iteration/cost/completed
  progress, stop hooks, context guard, token-budget trim, multimodal prep,
  streaming forwarder, native/text parse, circuit breaker, max-iter error).
- run_tool_call_loop is now a ~30-line adapter that builds a RegistryToolSource
  and hands off. Signature unchanged; bus/triage/summarizer/tests unaffected.

tool_loop + harness suites: 128/128 pass.
